refactored functions to exist in lib folder
